backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
sqlalchemy
torch
faster-whisper
langchain
langchain-openai
notion-client
mcp[fastmcp]
aiofiles
pydantic
pydantic-settings
